use createCacheKey also for parseFunction
