+++ src/utils/hash.js
@@ -1,0 +1,3 @@
+import { createHash } from 'crypto'
+
+export const md5sum = str => createHash('md5').update(str).digest('hex')
