+++ src/utils/index.js
@@ -2,8 +2,11 @@ import { BrowserError } from './errors'
 
 export * from './browser'
 export * from './compiler'
+export * from './cache'
 export * from './errors'
 export * from './fs'
+export * from './hash'
+export * from './package'
 export * from './page-functions'
 export * from './parser'
 export * from './timers'
