+++ src/utils/page-functions.js
@@ -1,17 +1,15 @@
-import { createHash } from 'crypto'
 import path from 'path'
 import webpack from 'webpack'
-import { camelCase } from '../utils'
 import {
   readFile,
   exists,
   stats,
   glob,
-  requireResolve
-} from './fs'
-
-let packageName
-let nodeModulesPath
+  camelCase,
+  findNodeModulesPath,
+  createCacheKey,
+  getCacheEntry
+} from '.'
 
 export async function createPageFunctions(page, sourceFiles, babelPresets) {
   const pageFunctions = {}
@@ -152,79 +150,4 @@ export function createWebpackConfig({
       ]
     }
   }
-}
-
-export function createCacheKey(fn, opts) {
-  if (typeof opts !== 'string') {
-    opts = JSON.stringify(opts)
-  }
-
-  return createHash('md5').update(`${fn}-x-${opts}`).digest('hex')
-}
-
-export async function getPackageName(modulesPath) {
-  if (!packageName) {
-    const { name } = await import(path.resolve(modulesPath, '..', 'package.json'))
-    packageName = name
-  }
-
-  return packageName
-}
-
-export async function getCacheEntry(filePath = '') {
-  const modulesPath = await findNodeModulesPath()
-  const name = await getPackageName(modulesPath)
-
-  return path.join(modulesPath, '.cache', name, filePath)
-}
-
-export async function checkNodeModulesPath(modulesPath) {
-  const pathExists = await exists(modulesPath)
-
-  if (pathExists) {
-    nodeModulesPath = modulesPath
-    return true
-  }
-
-  return false
-}
-
-export async function findNodeModulesPath(refresh) {
-  if (nodeModulesPath && !refresh) {
-    return nodeModulesPath
-  }
-
-  let modulesPath = path.resolve(__dirname, '../../..', 'node_modules')
-  if (await checkNodeModulesPath(modulesPath)) {
-    return nodeModulesPath
-  }
-
-  modulesPath = path.resolve('./node_modules')
-  if (await checkNodeModulesPath(modulesPath)) {
-    return nodeModulesPath
-  }
-
-  const hablePath = requireResolve('hable/package.json')
-  modulesPath = path.dirname(path.dirname(hablePath))
-  if (await checkNodeModulesPath(modulesPath)) {
-    return nodeModulesPath
-  }
-
-  modulesPath = __dirname
-  while (modulesPath.length > 1) {
-    const tryPath = path.join(modulesPath, 'node_modules')
-    modulesPath = path.dirname(modulesPath)
-
-    if (await checkNodeModulesPath(tryPath)) {
-      modulesPath = tryPath
-      break
-    }
-  }
-
-  if (modulesPath.length <= 1) {
-    modulesPath = ''
-  }
-
-  nodeModulesPath = modulesPath
-  return modulesPath
 }
